GraphQLの補完をできるように変更
