+++ .graphqlrc.yml
@@ -1,0 +1,2 @@
+schema: 'schema.json'
+documents: 'src/**/*.{graphql,js,ts,jsx,tsx}'
